+++ source/assets/locales/es.json
@@ -409,7 +409,11 @@
     "newChainType": "Nuevo Tipo de Cadena",
     "newNetworkUrl": "Nueva URL de Red",
     "newNetworkChainId": "Nuevo ChainID de Red",
-    "trustedWebsites": "SITIOS WEB DE CONFIANZA"
+    "trustedWebsites": "SITIOS WEB DE CONFIANZA",
+    "importWalletWarning": "IMPORT WALLET ADVERTENCIA!",
+    "thisActionErases": "Esta acción borra su billetera actual.",
+    "toRestoreThis": "Para restaurar esta billetera actual en el futuro, asegúrese de tener la frase de recuperación.",
+    "allYourSettings": "All your settings, like custom tokens and imported accounts, will get lost, however, your funds it’s saved!"
   },
   "start": {
     "paliDoesnt": "Pali no guarda tu contraseña. Recuperar el acceso requiere un reinicio de la billetera con tu Frase Secreta de Recuperación. Esta acción borra la billetera actual, las cuentas creadas, tokens personalizados y cuentas importadas de tu dispositivo, generando una nueva cuenta vinculada a la frase de reinicio.",
